// sdks/typescript/src/types.ts
export type UUID = string;
// ...
export interface RacewayContext {
  traceId: UUID;
  threadId: string;  // Unique ID for this request/async context
  parentId: UUID | null;
  rootId: UUID | null;
  clock: number;
  spanId: string;
  parentSpanId: string | null;
  distributed: boolean;
  clockVector: Array<[string, number]>;
  tracestate: string | null;
}